+++ demo/src/components/forms/mod.rs
@@ -1,2 +1,3 @@
 pub mod form_field_wrapper;
 pub mod form_state_debugger;
+pub mod form_content;
